# Don't use `alloca()` in `Object::emit_signalp()` to prevent stack overflow